+++ noodles-sam/src/header/program/builder.rs
@@ -1,4 +1,6 @@
-use std::collections::HashMap;
+//! SAM header program and fields.
+
+use std::{collections::HashMap, error, fmt};
 
 use super::{Program, Tag};
 
@@ -14,15 +16,34 @@ pub struct Builder {
     fields: HashMap<Tag, String>,
 }
 
+/// An error returned when a SAM header program fails to build.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum BuildError {
+    /// The ID is missing.
+    MissingId,
+}
+
+impl error::Error for BuildError {}
+
+impl fmt::Display for BuildError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::MissingId => f.write_str("missing ID"),
+        }
+    }
+}
+
 impl Builder {
     /// Sets a program ID.
     ///
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::Program;
-    /// let program = Program::builder().set_id("pg0").build();
+    /// let program = Program::builder().set_id("pg0").build()?;
     /// assert_eq!(program.id(), "pg0");
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
     pub fn set_id<I>(mut self, id: I) -> Self
     where
@@ -37,14 +58,16 @@ impl Builder {
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::Program;
     ///
     /// let program = Program::builder()
     ///     .set_id("pg0")
     ///     .set_name("noodles")
-    ///     .build();
+    ///     .build()?;
     ///
     /// assert_eq!(program.name(), Some("noodles"));
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
     pub fn set_name<I>(mut self, name: I) -> Self
     where
@@ -59,14 +82,16 @@ impl Builder {
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::Program;
     ///
     /// let program = Program::builder()
     ///     .set_id("pg0")
     ///     .set_command_line("cargo run")
-    ///     .build();
+    ///     .build()?;
     ///
     /// assert_eq!(program.command_line(), Some("cargo run"));
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
     pub fn set_command_line<I>(mut self, command_line: I) -> Self
     where
@@ -81,14 +106,16 @@ impl Builder {
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::Program;
     ///
     /// let program = Program::builder()
     ///     .set_id("pg1")
     ///     .set_previous_id("pg0")
-    ///     .build();
+    ///     .build()?;
     ///
     /// assert_eq!(program.previous_id(), Some("pg0"));
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
     pub fn set_previous_id<I>(mut self, previous_id: I) -> Self
     where
@@ -103,14 +130,16 @@ impl Builder {
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::Program;
     ///
     /// let program = Program::builder()
     ///     .set_id("pg0")
     ///     .set_description("noodles")
-    ///     .build();
+    ///     .build()?;
     ///
     /// assert_eq!(program.description(), Some("noodles"));
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
     pub fn set_description<I>(mut self, description: I) -> Self
     where
@@ -125,14 +154,16 @@ impl Builder {
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::Program;
     ///
     /// let program = Program::builder()
     ///     .set_id("pg0")
     ///     .set_version("0.1.0")
-    ///     .build();
+    ///     .build()?;
     ///
     /// assert_eq!(program.version(), Some("0.1.0"));
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
     pub fn set_version<I>(mut self, version: I) -> Self
     where
@@ -147,6 +178,7 @@ impl Builder {
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::{program::Tag, Program};
     ///
     /// let zn = Tag::Other(String::from("zn"));
@@ -154,9 +186,10 @@ impl Builder {
     /// let program = Program::builder()
     ///     .set_id("pg0")
     ///     .insert(zn.clone(), String::from("noodles"))
-    ///     .build();
+    ///     .build()?;
     ///
     /// assert_eq!(program.fields().get(&zn), Some(&String::from("noodles")));
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
     pub fn insert<I>(mut self, tag: Tag, value: I) -> Self
     where
@@ -171,18 +204,32 @@ impl Builder {
     /// # Examples
     ///
     /// ```
+    /// # use noodles_sam::header::program::builder;
     /// use noodles_sam::header::Program;
-    /// let program = Program::builder().set_id("rg0").build();
+    /// let program = Program::builder().set_id("rg0").build()?;
+    /// # Ok::<(), builder::BuildError>(())
     /// ```
-    pub fn build(self) -> Program {
-        Program {
-            id: self.id.expect("missing id"),
+    pub fn build(self) -> Result<Program, BuildError> {
+        let id = self.id.ok_or(BuildError::MissingId)?;
+
+        Ok(Program {
+            id,
             name: self.name,
             command_line: self.command_line,
             previous_id: self.previous_id,
             description: self.description,
             version: self.version,
             fields: self.fields,
-        }
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_build() {
+        assert_eq!(Builder::default().build(), Err(BuildError::MissingId));
     }
 }